+++ lib/ethpy/ethpy/hyperdrive/interface/__init__.py
@@ -1,3 +1,3 @@
 """High-level interface for the Hyperdrive market."""
-from .hyperdrive_read_interface import HyperdriveReadInterface
-from .hyperdrive_read_write_interface import HyperdriveReadWriteInterface
+from .read_interface import HyperdriveReadInterface
+from .read_write_interface import HyperdriveReadWriteInterface
