+++ lib/ethpy/ethpy/hyperdrive/interface/_block_getters.py
@@ -7,7 +7,7 @@ if TYPE_CHECKING:
     from eth_typing import BlockNumber
     from web3.types import BlockData, BlockIdentifier, Timestamp
 
-    from .hyperdrive_read_interface import HyperdriveReadInterface
+    from .read_interface import HyperdriveReadInterface
 
 
 def _get_block(interface: HyperdriveReadInterface, block_identifier: BlockIdentifier) -> BlockData:
